add tagging to articles model
this works partially. you can create a tag within the article form, and
tags are found or created as needed. however if the user adds two tags,
and the two separate tags are created, when the user returns to the edit
form, the two tags have been combined into one. thus on submitting the
form again, a new tag is created combining the original tags into one
which is not helpful. adding comments to issue #214 to clarify.