Trying to add rspec with stubbing for #331
Still not there. Moved all of these potential tests to
spec/features/external_requests_spec.rb

Added stubbing language from the error message to rails_helper.rb

We are past the stubbing error, and now need an appropriate results
matcher to confirm that the api was called correctly, which I think is
the best we can do if we are not going to actually hi the api.