Rebranching from master to implement MailChimp API
Still need to add tests which stub or mock the API returns. I have not
found a good tutorial for how to do this.  At this point, though the UX
is sub par without the fixes implemeted in this branch, I am moving on
to fixing the view to allow showing of recently updated cases.