for #491 retrieving tweets based on entered hashtag
With this commit article.tweets method retrieves the last 10 tweets based on the
first hashtag associated with an article. instance variable is set in
the controller. note: there may be too many unnecessary instance
variables in the show action.

need to figure out a way to accommodate multiple hashtags and create an
array of all associated tweets ordered by date so that they can be
displayed in order