for #511 redirect old url to updated path
In the past, when a slug-url was updated, the old url was invalid,
resulting in 404 errors for people and crawlers using old links. With this commit,
those old links will still work for the article.

The concern with this behavior is that having two urls for the sme page
could result in an SEO penalty for duplicate content. So the other part
of this commit, redirects the old link to the updated one.